requirements.txt
openai
pandas
psycopg[binary]
python-dotenv
